Code/requirements.txt
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==2.0.0
click==8.3.1
cssselect2==0.8.0
defusedxml==0.7.1
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
opencv-python==4.13.0.92
pillow==12.1.0
pycparser==3.0
pypng==0.20220715.0
svgwrite==1.4.3
tinycss2==1.5.1
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.0.1
zxing==1.0.3
zxing-cpp==2.3.0
